// crates/admin-app/src/lib/api/models/CreateItemRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateItemRequest = {
    class?: string | null;
    description?: string | null;
    equipmentSlot?: string | null;
    inventoryType: string;
    levelReq?: number | null;
    maxStack?: number | null;
    name: string;
    rarity: string;
    stats?: any;
};